# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI-Powered Supermarket Store Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: lowStock
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/barcode/{barcode}:
    get:
      operationId: getProductByBarcode
      tags: [products]
      summary: Get product by barcode
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  # INVENTORY
...
  /purchases:
    get:
      operationId: listPurchaseOrders
      tags: [purchases]
      summary: List purchase orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: supplierId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Purchase orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
    post:
      operationId: createPurchaseOrder
      tags: [purchases]
      summary: Create purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchases/{id}:
    get:
      operationId: getPurchaseOrder
      tags: [purchases]
      summary: Get purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
    patch:
      operationId: updatePurchaseOrder
      tags: [purchases]
      summary: Update purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchases/{id}/status:
    patch:
      operationId: updatePurchaseOrderStatus
      tags: [purchases]
      summary: Update purchase order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"

  # SALES
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, sku, costPrice, sellingPrice, unit]
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: string
        barcode:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        unit:
          type: string
        costPrice:
          type: number
        sellingPrice:
          type: number
        tax:
          type: number
        discount:
          type: number
        expiryDate:
          type: ["string", "null"]
        batchNumber:
          type: ["string", "null"]
        minStock:
          type: number
        maxStock:
          type: number
        reorderPoint:
          type: number
        currentStock:
          type: number
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, sku, costPrice, sellingPrice, unit]
      properties:
        name:
          type: string
        sku:
          type: string
        barcode:
          type: string
        description:
          type: string
        categoryId:
          type: ["integer", "null"]
        brand:
          type: string
        unit:
          type: string
        costPrice:
          type: number
        sellingPrice:
          type: number
        tax:
          type: number
        discount:
          type: number
        expiryDate:
          type: string
        batchNumber:
          type: string
        minStock:
          type: number
        maxStock:
          type: number
        reorderPoint:
          type: number
        initialStock:
          type: number
        imageUrl:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        sku:
          type: string
        barcode:
          type: string
        description:
          type: string
        categoryId:
          type: ["integer", "null"]
        brand:
          type: string
        unit:
          type: string
        costPrice:
          type: number
        sellingPrice:
          type: number
        tax:
          type: number
        discount:
          type: number
        expiryDate:
          type: string
        batchNumber:
          type: string
        minStock:
          type: number
        maxStock:
          type: number
        reorderPoint:
          type: number
        imageUrl:
          type: string
        status:
          type: string
          enum: [active, inactive]
    ProductList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    PurchaseOrderItem:
      type: object
      required: [productId, quantity, unitCost]
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        receivedQty:
          type: number
        unitCost:
          type: number
        total:
          type: number
    PurchaseOrder:
      type: object
      required: [id, supplierId, status, total, createdAt]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        supplierId:
          type: integer
        supplierName:
          type: string
        status:
          type: string
          enum: [draft, approved, ordered, received, completed, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItem"
        subtotal:
          type: number
        tax:
          type: number
        total:
          type: number
        notes:
          type: ["string", "null"]
        expectedDate:
          type: ["string", "null"]
        createdAt:
          type: string
    PurchaseOrderInput:
      type: object
      required: [supplierId, items]
      properties:
        supplierId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItem"
        notes:
          type: string
        expectedDate:
          type: string
    PurchaseOrderUpdate:
      type: object
      properties:
        supplierId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItem"
        notes:
          type: string
        expectedDate:
          type: string
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        notes:
          type: string
...